update some files

+++ SOURCE/Algorithms/header/bubbleSort.hpp
@@ -5,7 +5,7 @@
 #ifndef BUBBLE_SORT_HPP
 #define BUBBLE_SORT_HPP
 
-template <class T> void bubbleSort(std::vector<T> &arr, long long &count_comparison);
+template <class T> void bubbleSort(std::vector<T> &arr, size_t &count_comparison);
 template <class T> void bubbleSort(std::vector<T> &arr);
 
 #endif // BUBBLE_SORT_HPP
